Cleanup: merged ProtectedNodeContainer into VirtualMachine.
With the big simplification we did on the protect mechanism, it
was not worth it anymore to have it in a separate file.
